Moved to a base class for TCP statemachine to disconnect from netmap and allow testing. Added a base test for the statemachine
