+++ src/Magma.Transport.Tcp/Magma.Transport.Tcp.csproj
@@ -14,6 +14,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Magma.Common\Magma.Common.csproj" />
     <ProjectReference Include="..\Magma.Internet.Ip\Magma.Internet.Ip.csproj" />
+    <ProjectReference Include="..\Magma.Link\Magma.Link.csproj" />
   </ItemGroup>
 
 </Project>
